Edited gitignore

+++ TreeGenerator/Tree.cpp
@@ -24,7 +24,7 @@ void Tree::init() {
 	pushTriangle(&vertices, p1, p2, p3, 1.0f);
 	pushTriangle(&vertices, p1, p3, p4, 1.0f);
 	
-	branch(&vertices, position, rotation, rand() % 8 + 8, 12);
+	branch(&vertices, position, rotation, float(rand() % 8 + 8), 12.0f);
 
 	float *verticesArray = (float*) malloc(vertices.size() * sizeof(float));
 	std::copy(vertices.begin(), vertices.end(), verticesArray);
@@ -64,7 +64,7 @@ void Tree::branch(std::vector<float>* vertices, glm::fvec3 position, glm::fquat 
 
 		addCylinder(vertices, length, length / 8.0f);
 
-		float nbranches = rand() % numBranches + 2;
+		float nbranches = float(rand() % numBranches + 2);
 
 		float offsetAngle = (rand() % 100) / 100.0f * PI * 2.0f;
 		
@@ -75,7 +75,7 @@ void Tree::branch(std::vector<float>* vertices, glm::fvec3 position, glm::fquat 
 			branch(vertices, point1, rotation * glm::fquat(glm::fvec3(cosf(i / nbranches * PI * 2.0f + offsetAngle) * smallV, sinf(i / nbranches * PI * 2.0f + offsetAngle) * smallV, 0)), length * (0.7f + (rand() % 100) / 200.0f - 0.25f), depth - 1 - rand() % 3);
 		}
 
-		if(rand() % 100 > pineLike) branch(vertices, point1, rotation, length * 0.9, depth - 1);
+		if(rand() % 100 > pineLike) branch(vertices, point1, rotation, length * 0.9f, depth - 1);
 	}
 }
 void Tree::update(float deltaTime) {
@@ -86,7 +86,7 @@ void Tree::draw(unsigned int shader) {
 	glDrawArrays(GL_TRIANGLES, 0, vertexCount / 3);
 }
 void Tree::addCylinder(std::vector<float>* vertices, float height, float radius) {
-	int n = std::max(height * 3.0f, 3.0f);
+	int n = int(std::max(height * 3.0f, 3.0f));
 
 	glm::fvec3 c1(0, height / 2.0f, 0);
 	glm::fvec3 c2(0, -height / 2.0f, 0);
